:truck: rename: 폴더 이동
헤더와 푸터 컨테이너 폴더로 이동, 컨테이너에 메인 폴더 추가 후 파일들 메인으로 이동

Resolves: #3
