--- aif_group/src/app/constants/navigationTabs.ts
@@ -1,3 +1,0 @@
-// import { useRef } from "react";
-
-// const element = useRef(null);
