+++ aif_group/src/containers/main/AIImageSection.tsx
