+++ aif_group/src/containers/main/MainBanner.tsx
